backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
torch==2.1.0
torchvision==0.16.0
grad-cam==1.4.8
opencv-python-headless==4.8.0.76
Pillow==10.0.0
numpy==1.24.3
SQLAlchemy==2.0.21
pandas==2.0.3